projet_cnn_mission1_v7/projet_cnn/requirements.txt
tensorflow>=2.12.0
numpy>=1.23.0
matplotlib>=3.6.0
scikit-learn>=1.1.0
seaborn>=0.12.0
Pillow>=9.0.0
flask>=3.0.0
